# backend/OAS_HASKI_v1.0.yaml
openapi: "3.0.3"
info:
  description: "This is the HASKI API documentation"
  version: "1.0"
  title: "HASKI API"
servers:
  - url: http://haksi.whatever.com/v1.0
...
paths:
  /elements:
    get:
      tags:
      - "HASKI-intern"
      summary: "Get all elements"
      description: "Get the data for all Elements"
      operationId: "getAllElements"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Elements'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    post:
      tags:
      - "HASKI-intern"
      summary: "Upload element"
      description: "Upload the information of an element"
      operationId: "uploadElement"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementUpload'
      responses:
        "201":
          description: "Element was created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
  /elements/{elementId}:
    get:
      tags: 
      - "HASKI-intern"
      summary: "Get element"
      description: "Get the data for a specific Element"
      operationId: "getElementData"
      parameters:
      - name: "elementId"
        in: "path"
        description: "ID of the element"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    put:
      tags: 
      - "HASKI-intern"
      summary: "Update element"
      description: "Update the information of an element"
      operationId: "updateElement"
      parameters:
      - name: "elementId"
        in: "path"
        description: "ID of the element"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "body"
        in: "header"
        description: "The new information about an element"
        required: true
        schema: 
          $ref: "#/components/schemas/Element"
      responses:
        "201":
          description: "Element was updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
  /learningPath:
    post:
      tags:
      - "HASKI-intern"
      summary: "Upload learning path"
      description: "Upload the inital learning path"
      operationId: "uploadInitalLearningPath"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningPathUpload'
      responses:
        "201":
          description: "Learning Path was created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningPath"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
...
  /learningWay/{studentId}/{learningWayId}:
    get:
      tags: 
      - "HASKI-intern"
      summary: "Get Learning Way"
      description: "Get the currently saved Learning Way for a specific student"
      operationId: "getLearningWay"
      parameters:
      - name: "studentId"
        in: "path"
        description: "ID of the student"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "learningWayId"
        in: "path"
        description: "ID of the learning way"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningWay'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    put:
      tags: 
      - "HASKI-intern"
      summary: "Update Learning Way"
      description: "Update the Learning Way of a student by themself"
      operationId: "updateLearningWay"
      parameters:
      - name: "studentId"
        in: "path"
        description: "ID of the student"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "learningWayId"
        in: "path"
        description: "ID of the learning way"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "body"
        in: "header"
        description: "The new Learning Way"
        required: true
        schema: 
          $ref: "#/components/schemas/LearningWayUpload"
      responses:
        "201":
          description: "Learning Way was updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningWay"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
  /modules:
    get:
      tags:
      - "HASKI-intern"
      summary: "Get all modules"
      description: "Get the data for all Modules"
      operationId: "getAllModules"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Modules'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    post:
      tags:
      - "HASKI-intern"
      summary: "Upload module"
      description: "Upload the information of a module"
      operationId: "uploadModule"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpload'
      responses:
        "201":
          description: "Element was created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
  /modules/{moduleId}:
    get:
      tags: 
      - "HASKI-intern"
      summary: "Get module"
      description: "Get the data for a specific Module"
      operationId: "getModuleData"
      parameters:
      - name: "moduleId"
        in: "path"
        description: "ID of the module"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleUpload'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    put:
      tags: 
      - "HASKI-intern"
      summary: "Update module"
      description: "Update the information of an module"
      operationId: "updateModule"
      parameters:
      - name: "moduleId"
        in: "path"
        description: "ID of the module"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "body"
        in: "header"
        description: "The new information about an module"
        required: true
        schema: 
          $ref: "#/components/schemas/ModuleUpload"
      responses:
        "201":
          description: "Module was updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Module not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
...
  /student:
    get:
      tags: 
      - "HASKI-intern"
      summary: "Get all students"
      description: "Get the data for all Students"
      operationId: "getAllStudentData"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Students'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    post:
      tags:
      - "HASKI-intern"
      summary: "Upload student"
      description: "Upload the inital information of a student"
      operationId: "uploadInitalStudent"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpload'
      responses:
        "201":
          description: "Student was created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
  /student/{studentId}:
    get:
      tags: 
      - "HASKI-intern"
      summary: "Get student"
      description: "Get the data for a specific Student"
      operationId: "getStudentData"
      parameters:
      - name: "studentId"
        in: "path"
        description: "ID of the student"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
    put:
      tags: 
      - "HASKI-intern"
      summary: "Update student"
      description: "Update the student information and/or learning style"
      operationId: "updateStudent"
      parameters:
      - name: "studentId"
        in: "path"
        description: "ID of the student"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "body"
        in: "header"
        description: "The new Learning Style and Student Information"
        required: true
        schema: 
          $ref: "#/components/schemas/StudentUpdate"
      responses:
        "201":
          description: "Student was updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: "Invalid ID supplied"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
        "404":
          description: "Student not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralError"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    Element:
      type: "object"
      required:
      - "id"
      - "name"
      - "difficulty"
      - "creationDate"
      - "module"
      - "averageDuration"
      - "semester"
      - "style"
      - "proLIST"
      - "contraLIST"
      - "type"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1234
          description: "ID of the element"
        name: 
          type: "string"
          example: "My test element"
          description: "Name of the element"
        difficulty:
          type: "string"
          example: "Leicht"
          description: "Level of difficulty of the element"
        creationDate:
          type: "integer"
          format: "int64"
          example: 1656330943
          description: "Date element was created in UNIX time stamp"
        module:
          type: "string"
          example: "Informatik I"
          description: "Module the element belongs to"
        averageDuration:
          type: "integer"
          format: "int64"
          example: 180
          description: "The time it takes in average to complete element in minutes"
        semester: 
          type: "integer"
          format: "int64"
          example: 1
          description: "The semester, in which the element will be done normally"
        style:
          type: "string"
          example: "visual"
          description: "Format the elemnt has"
        type: 
          type: "string"
          example: "PDF"
        proLIST:
          type: "string"
          example: "Lorem Ipsum"
          description: "supported FSLSM-Dimension"
        contraLIST:
          type: "string"
          example: "Lorem Ipsum bad"
          description: "preventing FSLSM-Dimension"
    Elements:
      type: "object"
      properties:
        elements:
          type: "array"
          items: 
            $ref: '#/components/schemas/Element'
...
    ElementUpload:
      type: "object"
      required:
      - "name"
      - "date"
      - "topic"
      - "semester"
      - "style"
      - "type"
      properties:
        name:
          type: "string"
          example: "My test Element"
        date: 
          type: "string"
          format: "date-time"
        topic:
          type: "string"
          example: "Informatik I"
        semester:
          type: "integer"
          format: "int64"
          example: 1
        style: 
          type: "string"
          example: "visual"
        type: 
          type: "string"
          example: "PDF"
    GeneralError: 
      type: "object"
      required: 
      - "code"
      - "message"
      properties: 
        code: 
          type: "integer"
          format: "int32"
        message: 
          type: "string"
    LearningPath:
      type: "object"
      required: 
      - "id"
      - "moduleId"
      - "module"
      - "elements"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 123
          description: "ID of the learning path"
        moduleId:
          type: "integer"
          format: "int64"
          example: 987
          description: "ID of the module the path belongs to"
        module:
          type: "string"
          example: "Informatik I"
          description: "Name of the module the path belongs to"
        elements:
          type: "array"
          items: 
            $ref: '#/components/schemas/ElementMeta'
          description: "Elements belonging to the learning path"
    LearningPathUpload:
      type: "object"
      required: 
      - "moduleId"
      - "module"
      - "elements"
      properties:
        moduleId:
          type: "integer"
          format: "int64"
          example: 123
          description: "ID of the module"
        module:
          type: "string"
          example: "Informatik I"
          description: "Name of the module the path belongs to"
        elements:
          type: "array"
          items: 
            $ref: '#/components/schemas/ElementMeta'
          description: "Elements belonging to the learning path"
...
    ModuleUpload:
      type: "object"
      required: 
      - "name"
      - "module"
      - "semester"
      properties:
        name:
          type: "string"
          example: "Einführung in die Programmierung"
        module:
          type: "string"
          example: "IT-1234"
        semester:
          type: "integer"
          format: "int64"
          example: 1
    Modules:
      type: "object"
      properties:
        modules:
          type: "array"
          items: 
            $ref: '#/components/schemas/Module'
...
    Students:
      type: "object"
      properties:
        students:
          type: "array"
          items: 
            $ref: '#/components/schemas/Student'
...